+++ R/vroom_fwf.R
@@ -29,6 +29,7 @@ vroom_fwf <- function(file,
                       altrep_opts = deprecated(),
                       num_threads = vroom_threads(),
                       progress = vroom_progress(),
+                      show_col_specs = NULL,
                       .name_repair = "unique") {
 
   verify_fwf_positions(col_positions)
@@ -56,6 +57,8 @@ vroom_fwf <- function(file,
 
   col_select <- vroom_enquo(rlang::enquo(col_select))
 
+  has_spec <- !is.null(col_types)
+
   col_types <- as.col_spec(col_types)
 
   out <- vroom_fwf_(file, as.integer(col_positions$begin), as.integer(col_positions$end),
@@ -72,8 +75,8 @@ vroom_fwf <- function(file,
   out <- vroom_select(out, col_select, id)
   class(out) <- c("spec_tbl_df", class(out))
 
-  if (is.null(col_types)) {
-    show_spec_summary(out, locale = locale)
+  if (should_show_col_spec(has_spec, show_col_specs)) {
+    show_col_specs(out, locale)
   }
 
   out