+++ NEWS.md
@@ -1,5 +1,7 @@
 # vroom (development version)
 
+* New `vroom(show_col_specs=)` argument to more simply control showing of column specifications after parsing.
+
 * `vroom()` no longer inadvertently calls `.name_repair` functions twice (#310).
 
 * Fix an inadvertent performance regression when reading values (#309)